+++ a.js
@@ -1,0 +1,2 @@
+const data = require("data-store")(".carpe-diem");
+console.log(data.get("foo.bar.deep"));