+++ src/cli.ts
@@ -8,13 +8,13 @@ import { parseDuration } from "./utils";
 import * as chrono from "chrono-node";
 import * as figlet from "figlet";
 import * as chalk from "chalk";
-import * as CarpeDiemData from "./data";
+import * as data from "./data";
 import { getTaskOptionsInteractive } from "./task";
 
 //console.clear();
 //console.log(chalk.yellow(figlet.textSync("Carpe Diem", { horizontalLayout: "full" })));
 
-yargs
+const program = yargs
   .usage("Usage: i <command> [options]")
   .epilogue("for more information, find our manual at http://github.com/tianyishi2001/carpe-diem/")
   //.describe("A handy task recorder for people who don't make plans but do care about their efficiency.")
@@ -80,7 +80,7 @@ yargs
       }
       let entry = await iDid(parsed);
       console.log(entry);
-      CarpeDiemData.addLogEntry(entry);
+      data.addLogEntry(entry);
     },
   })
   .command({
@@ -89,11 +89,14 @@ yargs
     builder: (yargs) => yargs.positional("item", { describe: "log/tasks/locations/summary (currently only log implemented)" }),
     handler: (argv) => {
       switch (argv.item) {
+        case "path":
+          console.log(data.showPath());
+          break;
         case "log":
-          console.log(CarpeDiemData.getLog());
+          console.log(data.getLog());
           break;
         case "tasks":
-          console.log(CarpeDiemData.getTaskDict());
+          console.log(data.getTaskDict());
           break;
         default:
           console.log("not implemented yet");
@@ -109,7 +112,7 @@ yargs
       switch (argv.item) {
         case "tasks":
           let task = await getTaskOptionsInteractive();
-          CarpeDiemData.updateTasks(task);
+          data.updateTasks(task);
           break;
         default:
           console.log("not implemented yet");
@@ -126,4 +129,6 @@ yargs
   .example("i did", "I did something (interactive mode)")
   .example("i will --do my task", 'Now I will do "my task" and start timing')
   .example("i will --do my task --for 4 hrs 30 min", 'Now I will do "my task" for 4 hours and half')
-  .example("i will (do) my task for 4 hrs 30 min", 'Same as above, in natural language (the "do" word is optional) [NOT IMPLEMENTED YET]').argv;
+  .example("i will (do) my task for 4 hrs 30 min", 'Same as above, in natural language (the "do" word is optional) [NOT IMPLEMENTED YET]');
+
+program.argv;